refactor variable type spec
