+++ impl/nodejs/src/util/index.js
@@ -2,7 +2,6 @@ const SpecialCharacters = require('./special-characters');
 const error = require('./error');
 const arrayify = require('./arrayify');
 const metaData = require('./meta-data');
-const expression = require('./expression');
 const { required, checkType } = require('./validation');
 const { loadJSON } = require('./load-json');
 
@@ -11,7 +10,6 @@ module.exports = {
     error,
     arrayify,
     metaData,
-    expression,
     required,
     checkType,
     loadJSON
